compararImagem/requirements.txt
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
matplotlib>=3.7.0
seaborn>=0.12.0
